Move into another package to avoid circular deps
