+++ haystack/pipeline_utils/__init__.py
@@ -1,3 +1,4 @@
 from haystack.pipeline_utils.rag import build_rag_pipeline
+from haystack.pipeline_utils.indexing import build_indexing_pipeline
 
-__all__ = ["build_rag_pipeline"]
+__all__ = ["build_rag_pipeline", "build_indexing_pipeline"]
