+++ haystack/utils/__init__.py
@@ -1,6 +1,5 @@
 from haystack.utils.expit import expit
-from haystack.utils.requests_utils import request_with_retry
 from haystack.utils.filters import document_matches_filter
-from haystack.utils.indexing import build_indexing_pipeline
+from haystack.utils.requests_utils import request_with_retry
 
-__all__ = ["expit", "request_with_retry", "document_matches_filter", "build_indexing_pipeline"]
+__all__ = ["expit", "request_with_retry", "document_matches_filter"]
